Refactor FFmpeg and audio processing utilities before merging PR
- Remove redundant ffmpeg_wrapper.py and consolidate FFmpeg command handling in utils.py
- Update import statements and function references across multiple modules
- Add run_ffmpeg_command() to utils.py with thread management
- Remove unused audio processing functions from audio_alignment.py and audio_generation.py
- Update .pylintrc to allow more flexible linting rules
- Remove ffmpeg-python from requirements_core.txt
- Standardize log message handling with new LOG_TTV_DIR_CREATED constant
