+++ .vscode/settings.json
@@ -4,88 +4,4 @@
     ],
     "python.testing.unittestEnabled": false,
     "python.testing.pytestEnabled": true,
-    "cSpell.words": [
-        "acodec",
-        "acrossfade",
-        "addopts",
-        "adelay",
-        "afade",
-        "aformat",
-        "afplay",
-        "aloop",
-        "amix",
-        "anullsrc",
-        "aout",
-        "aresample",
-        "asetpts",
-        "asyncio",
-        "atempo",
-        "atrim",
-        "autoexit",
-        "autouse",
-        "boxcolor",
-        "Buildx",
-        "crossfade",
-        "dalle",
-        "deepskyblue",
-        "dejavu",
-        "direnv",
-        "dotenv",
-        "envrc",
-        "ffplay",
-        "filterwarnings",
-        "fltp",
-        "fontcolor",
-        "fontfile",
-        "fontsize",
-        "Hotword",
-        "hotwords",
-        "LANCZOS",
-        "lavfi",
-        "libmp",
-        "libsndfile",
-        "libx",
-        "loudnorm",
-        "moviepy",
-        "musicgen",
-        "musicout",
-        "nodisp",
-        "nokey",
-        "noprint",
-        "noto",
-        "numpy",
-        "opencv",
-        "Ozer's",
-        "portaudio",
-        "psutil",
-        "pyaudio",
-        "pydantic",
-        "pydub",
-        "pylint",
-        "pytest",
-        "PYTHONUNBUFFERED",
-        "pytree",
-        "reencode",
-        "reencoded",
-        "rlist",
-        "scikit",
-        "scipy",
-        "sounddevice",
-        "soundfile",
-        "STARTPTS",
-        "stillimage",
-        "Suno",
-        "Taylors",
-        "tcgetattr",
-        "TCSADRAIN",
-        "tcsetattr",
-        "testpaths",
-        "textbbox",
-        "texttospeech",
-        "torchaudio",
-        "ultrafast",
-        "urandom",
-        "videofile",
-        "xdist"
-    ]
 }
