+++ tests/smoke/test_simulated_ttv_pipeline.py
@@ -36,12 +36,13 @@ from tests.integration.test_helpers import (
     wait_for_completion,
     validate_segment_count,
     validate_background_music,
-    parse_test_logs
+    extract_test_logs
 )
 from utils import get_tempdir
 from ttv.log_messages import (
     LOG_BACKGROUND_MUSIC_SUCCESS,
-    LOG_BACKGROUND_MUSIC_FAILURE
+    LOG_BACKGROUND_MUSIC_FAILURE,
+    LOG_TTV_DIR_CREATED
 )
 
 logger = logging.getLogger(__name__)
@@ -78,7 +79,7 @@ def test_simulated_pipeline_execution():
         f.write(output)
 
     # Get the output directory from the output by getting everything after the : following `Created TTV directory` from the output
-    output_dir = output.split("Created TTV directory: ")[1].split("\n")[0]
+    output_dir = output.split(LOG_TTV_DIR_CREATED)[1].split("\n")[0]
     print(f"Detected output directory: {output_dir}")
     # Validate all segments are present
     validate_segment_count(output, SIMULATED_PIPELINE_CONFIG)
