+++ ttv/audio_alignment.py
@@ -387,236 +387,3 @@ def get_audio_duration(audio_file: str, thread_id: str = None) -> float:
             f"{thread_prefix}Error getting audio duration: {str(e)}"
         )
         raise
-
-def adjust_segment_timing(
-    segment_file: str,
-    target_duration: float,
-    thread_id: str = None
-) -> bool:
-    """Adjust audio segment timing to match target duration.
-    
-    Args:
-        segment_file: Path to the audio segment file
-        target_duration: Target duration in seconds
-        thread_id: Optional thread ID for logging
-        
-    Returns:
-        bool: True if adjustment was successful, False otherwise
-    """
-    try:
-        thread_prefix = f"{thread_id} " if thread_id else ""
-        current_duration = get_audio_duration(segment_file, thread_id)
-        
-        if abs(current_duration - target_duration) < 0.1:
-            Logger.print_info(
-                f"{thread_prefix}Duration already matches target "
-                f"({current_duration:.2f}s)"
-            )
-            return True
-
-        # Calculate tempo adjustment
-        tempo = (current_duration / target_duration) * 100
-        temp_output = f"{segment_file}.temp.mp3"
-
-        cmd = [
-            "ffmpeg", "-y",
-            "-i", segment_file,
-            "-filter:a", f"atempo={tempo/100}",
-            "-acodec", "libmp3lame",
-            "-q:a", "2",
-            temp_output
-        ]
-
-        Logger.print_info(
-            f"{thread_prefix}Adjusting tempo to {tempo:.1f}% "
-            f"(target duration: {target_duration:.2f}s)"
-        )
-        
-        result = subprocess.run(
-            cmd,
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-            check=True
-        )
-
-        # Replace original with adjusted version
-        os.replace(temp_output, segment_file)
-        
-        # Verify final duration
-        final_duration = get_audio_duration(segment_file, thread_id)
-        Logger.print_info(
-            f"{thread_prefix}Final duration: {final_duration:.2f}s"
-        )
-        
-        return abs(final_duration - target_duration) < 0.1
-
-    except subprocess.CalledProcessError as e:
-        Logger.print_error(
-            f"{thread_prefix}FFmpeg error: {e.stderr.decode()}"
-        )
-        return False
-    except (OSError, IOError) as e:
-        Logger.print_error(
-            f"{thread_prefix}File operation error: {str(e)}"
-        )
-        return False
-
-def process_audio_segments(
-    segments: List[str],
-    target_durations: List[float],
-    thread_id: str = None
-) -> bool:
-    """Process multiple audio segments to match target durations.
-    
-    Args:
-        segments: List of paths to audio segment files
-        target_durations: List of target durations in seconds
-        thread_id: Optional thread ID for logging
-        
-    Returns:
-        bool: True if all segments were processed successfully
-    """
-    if len(segments) != len(target_durations):
-        raise ValueError(
-            "Number of segments must match number of target durations"
-        )
-
-    thread_prefix = f"{thread_id} " if thread_id else ""
-    success = True
-
-    for i, (segment, duration) in enumerate(zip(segments, target_durations)):
-        Logger.print_info(
-            f"{thread_prefix}Processing segment {i+1}/{len(segments)}"
-        )
-        
-        if not adjust_segment_timing(segment, duration, thread_id):
-            Logger.print_error(
-                f"{thread_prefix}Failed to adjust timing for segment {i+1}"
-            )
-            success = False
-
-    return success
-
-def align_audio_with_text(
-    audio_file: str,
-    text: str,
-    output_file: str,
-    thread_id: str = None
-) -> bool:
-    """Align audio with text and save the result.
-    
-    Args:
-        audio_file: Path to the input audio file
-        text: Text to align with audio
-        output_file: Path to save the aligned audio
-        thread_id: Optional thread ID for logging
-        
-    Returns:
-        bool: True if alignment was successful
-    """
-    try:
-        thread_prefix = f"{thread_id} " if thread_id else ""
-        Logger.print_info(
-            f"{thread_prefix}Aligning audio with text: {text[:50]}..."
-        )
-
-        # Get word-level captions
-        captions = create_word_level_captions(
-            audio_file,
-            text,
-            thread_id=thread_id
-        )
-        if not captions:
-            return False
-
-        # Create segments based on captions
-        segments = []
-        for caption in captions:
-            segment = {
-                'start': caption.start_time,
-                'end': caption.end_time,
-                'text': caption.text
-            }
-            segments.append(segment)
-
-        # Create FFmpeg filter complex for segments
-        filter_complex = create_filter_complex(segments)
-        if not filter_complex:
-            return False
-
-        # Apply filter complex using FFmpeg
-        cmd = [
-            "ffmpeg", "-y",
-            "-i", audio_file,
-            "-filter_complex", filter_complex,
-            "-acodec", "libmp3lame",
-            "-q:a", "2",
-            output_file
-        ]
-
-        Logger.print_info(f"{thread_prefix}Applying audio alignment...")
-        result = subprocess.run(
-            cmd,
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-            check=True
-        )
-
-        Logger.print_info(
-            f"{thread_prefix}Successfully aligned audio: {output_file}"
-        )
-        return True
-
-    except subprocess.CalledProcessError as e:
-        Logger.print_error(
-            f"{thread_prefix}FFmpeg error: {e.stderr.decode()}"
-        )
-        return False
-    except (OSError, IOError) as e:
-        Logger.print_error(
-            f"{thread_prefix}File operation error: {str(e)}"
-        )
-        return False
-    except Exception as e:
-        Logger.print_error(
-            f"{thread_prefix}Unexpected error in align_audio_with_text: {str(e)}"
-        )
-        Logger.print_error(f"Traceback: {traceback.format_exc()}")
-        return False
-
-def create_filter_complex(segments: List[dict]) -> str:
-    """Create FFmpeg filter complex string for audio segments.
-    
-    Args:
-        segments: List of segment dictionaries with start, end, and text
-        
-    Returns:
-        str: FFmpeg filter complex string
-    """
-    try:
-        parts = []
-        for i, segment in enumerate(segments):
-            start = segment['start']
-            end = segment['end']
-            duration = end - start
-            
-            # Create segment filter
-            part = (
-                f"[0:a]atrim=start={start}:end={end},"
-                f"asetpts=PTS-STARTPTS[s{i}]"
-            )
-            parts.append(part)
-
-        # Concatenate all segments
-        parts.append(
-            "".join(f"[s{i}]" for i in range(len(segments))) +
-            f"concat=n={len(segments)}:v=0:a=1[out]"
-        )
-
-        return ";".join(parts)
-
-    except Exception as e:
-        Logger.print_error(f"Error creating filter complex: {str(e)}")
-        return ""
-
-# ... continue with other functions ...
