+++ ttv/audio_generation.py
@@ -193,67 +193,3 @@ def mix_audio_tracks(
             f"{thread_prefix}Error mixing audio tracks: {str(e)}"
         )
         return None
-
-def process_audio(
-    input_path: str,
-    output_path: str,
-    normalize: bool = True,
-    start_time: Optional[float] = None,
-    end_time: Optional[float] = None,
-    thread_id: Optional[str] = None
-) -> Optional[str]:
-    """Process an audio file with various operations.
-    
-    Args:
-        input_path: Path to input audio file
-        output_path: Path to save processed audio
-        normalize: Whether to normalize audio levels
-        start_time: Optional start time for trimming
-        end_time: Optional end time for trimming
-        thread_id: Optional thread ID for logging
-        
-    Returns:
-        Optional[str]: Path to processed audio if successful, None otherwise
-    """
-    thread_prefix = f"{thread_id} " if thread_id else ""
-    
-    try:
-        # Build ffmpeg command
-        command = ["ffmpeg", "-y", "-i", input_path]
-        
-        # Add filters
-        filters = []
-        
-        if normalize:
-            filters.append("loudnorm")
-            
-        if start_time is not None or end_time is not None:
-            trim_args = []
-            if start_time is not None:
-                trim_args.append(f"start={start_time}")
-            if end_time is not None:
-                trim_args.append(f"end={end_time}")
-            filters.append(f"atrim={':'.join(trim_args)}")
-            
-        if filters:
-            command.extend(["-af", ",".join(filters)])
-            
-        command.append(output_path)
-        
-        # Run ffmpeg
-        subprocess.run(
-            command,
-            check=True,
-            capture_output=True
-        )
-        
-        Logger.print_info(
-            f"{thread_prefix}Processed audio saved to: {output_path}"
-        )
-        return output_path
-        
-    except subprocess.CalledProcessError as e:
-        Logger.print_error(
-            f"{thread_prefix}Error processing audio: {str(e)}"
-        )
-        return None
