+++ ttv/log_messages.py
@@ -1,5 +1,8 @@
 """Constants for log messages used throughout the TTV pipeline."""
 
+# Directory creation messages
+LOG_TTV_DIR_CREATED = "Created TTV directory: "
+
 # Video segment creation and processing
 LOG_VIDEO_SEGMENT_CREATE = "Creating video segment: output_path"
 LOG_FFPROBE_COMMAND = "Running ffprobe command: ffprobe -v error -show_entries format=duration -of default=noprint_wrappers=1:nokey=1"
