+++ utils.py
@@ -21,6 +21,8 @@ import time
 import random
 from logger import Logger
 import queue
+from ttv.log_messages import LOG_TTV_DIR_CREATED
+import subprocess
 
 openai.api_key = os.environ.get("OPENAI_API_KEY")
 
@@ -38,6 +40,7 @@ def get_tempdir():
     if temp_dir == None:
         temp_dir = os.path.join(tempfile.gettempdir(), 'GANGLIA')
     os.makedirs(temp_dir, exist_ok=True)
+    Logger.print_info(f"{LOG_TTV_DIR_CREATED}{temp_dir}")
     return temp_dir
 
 def get_timestamped_ttv_dir() -> str:
@@ -58,13 +61,45 @@ def get_timestamped_ttv_dir() -> str:
         timestamp = datetime.now().strftime("%Y-%m-%d-%H-%M-%S")
         _CURRENT_TTV_DIR = os.path.join(get_tempdir(), "ttv", timestamp)
         os.makedirs(_CURRENT_TTV_DIR, exist_ok=True)
-        Logger.print_info(f"Created TTV directory: {_CURRENT_TTV_DIR}")
+        Logger.print_info(f"{LOG_TTV_DIR_CREATED}{_CURRENT_TTV_DIR}")
     
     return _CURRENT_TTV_DIR
 
 def get_config_path():
     """Get the path to the config directory relative to the project root."""
     return os.path.join(os.path.dirname(os.path.dirname(__file__)), 'config', 'ganglia_config.json')
+
+def run_ffmpeg_command(ffmpeg_cmd):
+    """
+    Run an FFmpeg command with managed thread allocation.
+    
+    Args:
+        ffmpeg_cmd: List of command arguments for FFmpeg
+    
+    Returns:
+        subprocess.CompletedProcess or None if the command fails
+    """
+    try:
+        # Use thread manager as context manager to track active operations
+        with ffmpeg_thread_manager as mgr:
+            # Get optimal thread count for this operation
+            thread_count = mgr.get_threads_for_operation()
+
+            # Insert thread count argument right after ffmpeg command
+            # Make a copy of the command to avoid modifying the original
+            cmd = ffmpeg_cmd.copy()
+            cmd.insert(1, "-threads")
+            cmd.insert(2, str(thread_count))
+
+            Logger.print_info(f"Running ffmpeg command with {thread_count} threads: {' '.join(cmd)}")
+            result = subprocess.run(cmd, check=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+            Logger.print_info(f"ffmpeg output: {result.stdout.decode('utf-8')}")
+            return result
+
+    except subprocess.CalledProcessError as e:
+        Logger.print_error(f"ffmpeg failed with error: {e.stderr.decode('utf-8')}")
+        Logger.print_error(f"ffmpeg command was: {' '.join(ffmpeg_cmd)}")
+        return None
 
 @lru_cache(maxsize=1)
 def get_system_info():
